fix bug

+++ plugin-echoprobe/EchoProbeParameters.h
@@ -31,14 +31,14 @@ public:
     std::optional<uint8_t> bw;
     std::optional<bool> sgi;
 
-    std::optional<int64_t> cf_begin;
-    std::optional<int64_t> cf_end;
-    std::optional<int64_t> cf_step;
+    std::optional<double> cf_begin;
+    std::optional<double> cf_end;
+    std::optional<double> cf_step;
     std::optional<uint32_t> cf_repeat;
 
-    std::optional<int32_t> sf_begin;
-    std::optional<int32_t> sf_end;
-    std::optional<int32_t> sf_step;
+    std::optional<double> sf_begin;
+    std::optional<double> sf_end;
+    std::optional<double> sf_step;
 
     uint32_t tx_max_retry;
     std::optional<uint8_t> ack_mcs;
